// src/api/back/models/UserProfileDecorationType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum UserProfileDecorationType {
    HAT = 'HAT',
    TITLE = 'TITLE',
    CHATICON = 'CHAT_ICON',
    CHATICONANIMATION = 'CHAT_ICON_ANIMATION'
}

export function UserProfileDecorationTypeFromJSON(json: any): UserProfileDecorationType {
    return UserProfileDecorationTypeFromJSONTyped(json, false);
}

export function UserProfileDecorationTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserProfileDecorationType {
    return json as UserProfileDecorationType;
}

export function UserProfileDecorationTypeToJSON(value?: UserProfileDecorationType | null): any {
    return value as any;
}